// app/models/responses/activity.ts
export interface AddActivityResponse {
  id: string;
  success: boolean;
}

export interface AddExpenseResponse {
  id: string;
  success: boolean;
}

export interface AddDebtResponse {
  id: string;
  success: boolean;
}

export interface AddParticipantResponse {
  id: string;
  success: boolean;
}

export interface DeleteActivityResponse {
  id: string;
  success: boolean;
}

export interface DeleteDebtResponse {
  id: string;
  success: boolean;
}

export interface DeleteExpenseResponse {
  id: string;
  success: boolean;
}

export interface DeleteParticipantResponse {
  id: string;
  success: boolean;
}